Fix for donate radio options
- Replaced <Input/> with DonateRadioButton component
- Renamed DonateRadioButton to DonateRadioOptions, plural name makes it clearer it renders multiple things
- Use priceId instead of amount for the form value name, which I think is what will actually be desired later
